grpc: moving files to extensions
