+++ test/common/config/BUILD
@@ -32,69 +32,6 @@ envoy_cc_test(
 )
 
 envoy_cc_test(
-    name = "delta_subscription_impl_test",
-    srcs = ["delta_subscription_impl_test.cc"],
-    deps = [
-        ":delta_subscription_test_harness",
-        "//envoy/config:xds_config_tracker_interface",
-        "//source/common/config:api_version_lib",
-        "//source/common/stats:isolated_store_lib",
-        "//source/extensions/config_subscription/grpc:grpc_subscription_lib",
-        "//test/mocks:common_lib",
-        "//test/mocks/config:config_mocks",
-        "//test/mocks/event:event_mocks",
-        "//test/mocks/grpc:grpc_mocks",
-        "//test/mocks/local_info:local_info_mocks",
-        "//test/mocks/runtime:runtime_mocks",
-        "//test/test_common:logging_lib",
-        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/endpoint/v3:pkg_cc_proto",
-        "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_test(
-    name = "delta_subscription_state_test",
-    srcs = ["delta_subscription_state_test.cc"],
-    deps = [
-        "//source/common/config:delta_subscription_state_lib",
-        "//source/common/config/xds_mux:delta_subscription_state_lib",
-        "//source/common/stats:isolated_store_lib",
-        "//source/extensions/config_subscription/grpc:grpc_subscription_lib",
-        "//test/mocks:common_lib",
-        "//test/mocks/config:config_mocks",
-        "//test/mocks/event:event_mocks",
-        "//test/mocks/grpc:grpc_mocks",
-        "//test/mocks/local_info:local_info_mocks",
-        "//test/mocks/runtime:runtime_mocks",
-        "//test/test_common:logging_lib",
-        "//test/test_common:test_runtime_lib",
-        "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
-        "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_test(
-    name = "delta_subscription_state_old_test",
-    srcs = ["delta_subscription_state_old_test.cc"],
-    deps = [
-        "//source/common/config:delta_subscription_state_lib",
-        "//source/common/stats:isolated_store_lib",
-        "//source/extensions/config_subscription/grpc:grpc_subscription_lib",
-        "//test/mocks:common_lib",
-        "//test/mocks/config:config_mocks",
-        "//test/mocks/event:event_mocks",
-        "//test/mocks/grpc:grpc_mocks",
-        "//test/mocks/local_info:local_info_mocks",
-        "//test/mocks/runtime:runtime_mocks",
-        "//test/test_common:logging_lib",
-        "//test/test_common:test_runtime_lib",
-        "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
-        "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_test(
     name = "sotw_subscription_state_test",
     srcs = ["sotw_subscription_state_test.cc"],
     external_deps = ["abseil_flat_hash_set"],
@@ -185,26 +122,6 @@ envoy_cc_test_library(
         "//test/mocks/upstream:cluster_manager_mocks",
         "//test/test_common:resources_lib",
         "//test/test_common:utility_lib",
-        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/endpoint/v3:pkg_cc_proto",
-        "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_test_library(
-    name = "delta_subscription_test_harness",
-    hdrs = ["delta_subscription_test_harness.h"],
-    deps = [
-        ":subscription_test_harness",
-        "//envoy/config:xds_config_tracker_interface",
-        "//source/common/common:utility_lib",
-        "//source/common/grpc:common_lib",
-        "//test/mocks/config:config_mocks",
-        "//test/mocks/config:custom_config_validators_mocks",
-        "//test/mocks/event:event_mocks",
-        "//test/mocks/grpc:grpc_mocks",
-        "//test/mocks/local_info:local_info_mocks",
-        "//test/mocks/runtime:runtime_mocks",
         "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
         "@envoy_api//envoy/config/endpoint/v3:pkg_cc_proto",
         "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
