+++ test/extensions/config_subscription/common/BUILD
@@ -43,10 +43,10 @@ envoy_cc_test(
     name = "subscription_impl_test",
     srcs = ["subscription_impl_test.cc"],
     deps = [
-        "//test/common/config:delta_subscription_test_harness",
         "//test/common/config:grpc_subscription_test_harness",
         "//test/common/config:subscription_test_harness",
         "//test/extensions/config_subscription/filesystem:filesystem_subscription_test_harness",
+        "//test/extensions/config_subscription/grpc:delta_subscription_test_harness",
         "//test/extensions/config_subscription/rest:http_subscription_test_harness",
     ],
 )
