+++ test/extensions/config_subscription/common/subscription_impl_test.cc
@@ -1,10 +1,10 @@
 #include <chrono>
 #include <memory>
 
-#include "test/common/config/delta_subscription_test_harness.h"
 #include "test/common/config/grpc_subscription_test_harness.h"
 #include "test/common/config/subscription_test_harness.h"
 #include "test/extensions/config_subscription/filesystem/filesystem_subscription_test_harness.h"
+#include "test/extensions/config_subscription/grpc/delta_subscription_test_harness.h"
 #include "test/extensions/config_subscription/rest/http_subscription_test_harness.h"
 
 using testing::InSequence;
